+++ src/main/java/KT/PLIP/population/dto/TimeBasedPopulationDto.java
@@ -13,9 +13,9 @@ public class TimeBasedPopulationDto {
         private String timeZone;        // 시간대 (새벽, 오전, 오후, 저녁, 밤)
         private String timeRange;       // 시간 범위 (예: "00:00-01:00")
         private String tmzonPdSe;       // 원본 시간대 코드
-        private Integer totalPopulation;
-        private Integer localPopulation;
-        private Integer tempForeignerPopulation;
-        private Integer longForeignerPopulation;
+        private Double totalPopulation;
+        private Double localPopulation;
+        private Double tempForeignerPopulation;
+        private Double longForeignerPopulation;
     }
 }
