+++ src/main/java/KT/PLIP/population/dto/WeekdayWeekendPopulationDto.java
@@ -7,8 +7,8 @@ import java.util.Map;
 @Getter
 @Setter
 public class WeekdayWeekendPopulationDto {
-    private Integer weekdayPopulation;
-    private Integer weekendPopulation;
-    private Map<String, Integer> ageGroupWeekday; // 연령대별 주중 인구
-    private Map<String, Integer> ageGroupWeekend; // 연령대별 주말 인구
+    private Double weekdayPopulation;
+    private Double weekendPopulation;
+    private Map<String, Double> ageGroupWeekday; // 연령대별 주중 인구
+    private Map<String, Double> ageGroupWeekend; // 연령대별 주말 인구
 }
